+++ Native/OpenDDSWrapper/ParticipantService.h
@@ -38,7 +38,7 @@ EXTERN_METHOD_EXPORT
 bool ParticipantService_GetIsShutdown();
 
 EXTERN_METHOD_EXPORT
-long ParticipantService_GetScheduler();
+int ParticipantService_GetScheduler();
 
 EXTERN_METHOD_EXPORT
 void ParticipantService_Scheduler(int scheduler);
